third_party/blink/web_tests/editing/pasteboard/copy-paste-first-line-in-textarea-expected.txt

PASS selectFirstLineIncludingLineBreak(); copyPaste(); textarea.selectionEnd is offsetAfterLineBreak
PASS successfullyParsed is true

TEST COMPLETE
